CM1705 - GA: Fundamentals of Programming
You are required to write a simple "Hello World" program which you should then amend to request the user's name (which is then used in the greeting output). The changes should be done on GitHub by first creating a public repository for the program and then creating a branch for the changes. The change should be committed, and the versions merged. For the purposes of this exercise, do not delete the branch.